services/requirements.txt
# FastAPI and server
# Updated versions for Python 3.14 compatibility
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# HTTP client for API calls
httpx>=0.25.0

# Data validation
# Using newer pydantic for Python 3.14 compatibility (pre-built wheels available)
pydantic>=2.5.0

# File handling
python-multipart>=0.0.6
# Pillow has build issues with Python 3.14 - skipping for now
# pillow>=10.1.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1

# Deepgram SDK for ASR/TTS
# Using v5.x API (code has been updated to match)
deepgram-sdk>=5.3.0


# Environment variables
python-dotenv>=1.0.0
